+++ FastAPI/services/product.py
@@ -1,4 +1,5 @@
 from models.product import Product
+from models.order import OrderProduct
 from sqlalchemy.orm import Session
 from dto import product
 
@@ -29,6 +30,15 @@ def update(data: product.Product, db: Session, id: int):
     return product
 
 def remove(db: Session, id: int):
+    order = db.query(OrderProduct).filter(
+        OrderProduct.product_id == id
+    ).first()
+
+    print(order)
+
+    if (order):
+        return {'error':'Нельзя удалить этот товар, т.к. с ним есть оформленный заказ. Для удаления этого товара сначала удалите все содержащие его заказы.'}
+
     product = db.query(Product).filter(Product.id==id).delete()
     db.commit()
     return product
